mqtt/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.4
django-ninja==1.3.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
influxdb-client==1.48.0
paho-mqtt==2.1.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.1
